fix(dsv41): fp8 GEMM column guards + f32 activation scales; add its missing GPU numerics test
Two real defects found while verifying the last two operator stubs:

1. NO COLUMN BOUNDS in the dense fp8 GEMM. gemm_fp8_kernel wrote the epilogue
   unconditionally and read the B operand / weight scales by raw column index,
   so any n that is not a multiple of the 64-column block tile corrupted memory
   and produced wrong results. The model's real shapes (4096/5120/2304/1280) are
   all multiples of 64, which is why it hid. Found by the new
   tests_dsv41_gemm_fp8.cu case n=96 (rel error 0.83). Fixed by clamping the
   B-row and scale column indices and guarding every epilogue store.
2. INCONSISTENT SCALE TYPES. dsv41_quant_fp8 writes f32 activation scales (the
   reference's scales_a is fp32 holding a power of two) while the GEMM read them
   as ue8m0 BYTES -- silently zeroing the output. Unified on f32 activation
   scales in the kernel, the ABI and the compressor's own quantiser + call
   sites; weight scales stay ue8m0 bytes (as in the checkpoint).

Also adds the dense GEMM's missing GPU numerics test (the most-used kernel in
the model had none), and extends the compile-check script to cover the
self-test programs.

Independently verified (re-ran all three on the GPU myself):
  tests_tcgen05_mxf4   (fp4 expert GEMM)      -> all cases EXACT
  tests_dsv41_attn     (indexer+compressor)   -> all cases passed
  tests_dsv41_gemm_fp8 (dense fp8 GEMM)       -> all cases OK (5/5, incl n=96)
Three .cu files link into one .so: 16 exported dsv41_* symbols.

+++ crates/ferrite-dsv41/src/kernels.rs
@@ -59,7 +59,9 @@ extern "C" {
     /// `bias` may be null. Uses m16n8k32 fp8 MMA with the epilogue scheme above.
     pub fn dsv41_gemm_fp8_mx(
         a: *const u8,
-        a_scale: *const u8,
+        // a_scale: f32 power-of-two activation scales, one per (row, k/32) --
+        // the reference's `scales_a` type. (w_scale below is ue8m0 BYTES.)
+        a_scale: *const f32,
         w: *const u8,
         w_scale: *const u8,
         bias: *const f32,
